Make scripts faster

+++ vpn-down.sh
@@ -4,10 +4,16 @@ set -euo pipefail
 
 IFS=$'\n\t'
 
+if [ ! -f ".vmname" ]; then
+  echo "VM not running?"
+  exit 1
+fi
+
 SSHUTTLEPID=$(pgrep sshuttle)
 if [ -n "$SSHUTTLEPID" ]; then
   echo "Stopping running sshuttle instance with PID $SSHUTTLEPID"
   kill "$SSHUTTLEPID"
 fi
 
-vagrant ssh --command "/vagrant/connect.sh down"
+source .vmname
+ssh -F .ssh_config_vagrant "$VMNAME" "/vagrant/connect.sh down"
